Overhaul of `scala.scalanative.unsafe.{stackalloc,alloc}` (#3411)
* Restore `stackalloc[T](Int)` and `alloc[T](Int)` methods. The now have special handling for literals integers and detect zero and negative literals at compile time. Non-literal values now have a runtime check to ensure that provide integer is positive integer. Unsigned variants are still present and have the same sementics as before. 
* `Intrinsic.stackalloc`signature has changed, methods take now generic `T` type used to resolve size of required memory. The `size: RawSize` argument no longer describes size of the allocated memory, instead it describe number of elements of sizeOf[T] to allocate. This change allows to skip some arithmetic operations at runtime/optimizer and allows to use the `nir.Type` argument of `nir.Op.Stackalloc` (previously always Byte). This work is based on #3260 
* 0-initialization of stackallocated memory is now moved to `Lowering` stage. Thanks to having a full information about allocated type, we can calculate exactly the required amount of memory at linktime after optimizer run, and allows to add optimizer runs designed to removal of zero-initialization of manually initialized memory. 
* All special intrinsic operations used as intermiediete step of compilation were moved to `Intrinsics.internal` object
* Removed most of no longer required `toUnsigned` conversion

+++ javalib/src/main/scala/java/lang/IEEE754Helpers.scala
@@ -67,7 +67,7 @@ private[java] object IEEE754Helpers {
     if (nChars == 0)
       throw new NumberFormatException(exceptionMsg(s))
 
-    val cStr: CString = stackalloc[scala.Byte]((nChars + 1).toUInt)
+    val cStr: CString = stackalloc[scala.Byte](nChars + 1)
 
     if (_numericCharSeqToCString(s, nChars, cStr) == false) {
       throw new NumberFormatException(exceptionMsg(s))
